--- algosdk/future/__init__.py
@@ -1,3 +1,0 @@
-from . import transaction
-
-name = "future"
