--- docs/algosdk/future/index.rst
@@ -1,7 +1,0 @@
-future
-======
-.. toctree::
-   :maxdepth: 10
-
-   transaction
-
