+++ examples/example.py
@@ -1,5 +1,5 @@
 from algosdk import encoding
-from algosdk.future import transaction
+from algosdk import transaction
 from algosdk import kmd
 from algosdk.v2client import algod
 from algosdk import account
