+++ tests/steps/application_v2_steps.py
@@ -6,10 +6,15 @@ import time
 import pytest
 from behave import given, step, then, when
 
-from algosdk import abi, atomic_transaction_composer, encoding, mnemonic
+from algosdk import (
+    abi,
+    atomic_transaction_composer,
+    encoding,
+    mnemonic,
+    transaction,
+)
 from algosdk.abi.contract import NetworkInfo
 from algosdk.error import ABITypeError, AtomicTransactionComposerError
-from algosdk.future import transaction
 from tests.steps.other_v2_steps import read_program
