+++ tests/steps/steps.py
@@ -15,8 +15,8 @@ from algosdk import (
     mnemonic,
     util,
     wallet,
+    transaction,
 )
-from algosdk.future import transaction
 
 
 @parse.with_pattern(r".*")
